Restructuring

+++ src/utils/file_utils.py
@@ -42,7 +42,7 @@ def get_start_end_time(time: str):
     else:
         start_time = datetime(settings.STARTYEAR + settings.RESTART * settings.BACKWARD_MULT, 1, 1, 0, 0)
     end_time = datetime(settings.STARTYEAR + (settings.RESTART + 1) * settings.BACKWARD_MULT, 1, 1, 0, 0)
-    simulation_length = np.abs((end_time - start_time).days)
+    simulation_length = np.abs((end_time - start_time).days).astype(int)
     if time == 'start':
         return start_time
     elif time == 'end':
